Stuck on SAXParser import error w/ Burp
Running this in burp produces an error related to importing SAXParser
when get_role/create_role is called.

This seems to be a long standing issue and I haven't been able to
workaround it yet.

Going to put the burp plugin on hold and work on making this a cli/
library. If we can figure out this later we'll be able to reuse
the library in the burp plugin though.